Refactors code to use chain::Listen and SpvClient
implements chain::Listen in all the watching components (Gatekeeper, Watcher, and Responder) to receive block data instead of using the receiver end of a broadcast channel.

ChainMonitor uses SpvClient to feed block data to the later instead of using the transmitter of the broadcast channel.

Side note: The Gatekeeper may not need the implement Listen, since the Watcher could feed him data once received. Leaving it as is makes it so the design is more modular and both components can be run in different devices. Same may apply to the Responder.