Replaces httpmock with jsonrpc and creates a custom BitcoindMock
httpmock was quite limited because the RPC nonces could not be properly set, so the library testing had several fixmes related to it.

This commit removes httpmock for now and replaces it with jsonrpc. Using the later
a pretty basic BitcoindMock has been created to be used in unittesting.

This also fixes Carrier::send_transaction for the case where the transaction was already on chain (RPC_VERIFY_ALREADY_IN_CHAIN) which was missing in the previous commit.