Removes AppointmentStatus from ExtendedAppointment
AppointmentStatus was meant to specify the status in which the appointment was found at. However, that info does not make sense within ExtendedAppointent since, when pulling appointment info, either an Appointment or a TransactionTracker will be returned. Therefore, the state may either be implicit or be part of a wrapping structure