Gets rid of get_blockchain_info, adds get_best_block_hash_and_height
`bitcoin_cli::get_blockchain_info` was only used to try the connection with bitcoind when `BitcoindClient` was being created. That involved, however, having a custom parser for the response that was not used elsewhere, plus having to maintain a custom `convert.rs` only for that.

In order to simplify the codebase, this commit deletes `get_blockchain_info` and all the stuff surrounding it and replaces it by `get_best_block_hash_and_height` which is used for the same purpose but can be parsed using `TryInto<(BlockHash, Option<u32>)> for JsonResponse` from `lightning-block-sync::convert`.