Arc + Mutex to Gatekeeper, Watcher and Responder
Uses `Arc<T>` instead of `&T` when passing instances to the `Gatekeeper` and
`Responder`. This is required so the `Watcher` can implement the `'static` trait,
given it is necessary for the `InternalAPI` implementation of `TowerServices`.

Also renames `borrowed` variables to give them a better naming and takes
lock acquiring out of loop whenever possible.