Removes outdated_users_cache from the gatekeeper
The purpose of the cache was to allow the Watcher and the Responder
to query data from the Gatekeeper during data deletion.  However, the
cache is actually not necessary. Chain listeners are called in order by the
ChainMonitor, and the Gatekeeper is placed last in main.rs. Therefore,
all data has been requested by the time the Gatekeeper needs to delete it.