Fixes appointment deletion when passed to the responder
Appointments were not being properly deleted when passed from the Watcher
to the Responder. Wether the Responder accepted the job or not was not
being checked, meaning the Appointment was being deleted as if it was
invalid.

This mainly means that once an Appointment was passed to the Responder it
was not only being deleted from memory, but also from the database. This
produced a cascade deletion, so the entry was not even in the Trackers table.

Adds deleted_appointment_from_memory to the Watcher and checks the DeliveryReceipt
when handling an appointment to the Responder. Also adds missing docs.