Cleans DBM. Adds batch remove methods and removes unused ones.
Some of the DBM methods were unused (or only used in tests). Moves/removes
the unnecessary methods and creates/updates batch deletion methods so
data can be removed in batch, avoiding partial deletion in case of a system
crash.

`DBM::batch_remove_appointments` is repurposed so it updates the  users
records in the same batch to also prevent partial update/deletion.