Updates appointment/tracker deletion methods.
For the Gatekeeper, `delete_appointments` (`delete_appointments_from_memory` now)
does not update the database anymore. The user update is performed by `DBM::batch_delete_appointments`
which is, in turn,called by the Watcher or the Responder during appointments/tracker cleanup.
The purpose of this is to make sure that partial data deletion cannot happen
(e.g. records are deleted from the appointments/trackers  tables but not
from users due to a system crash during the cleanup process).

For the Responder, data deletion is split in two methods, one to delete
data from memory only and another to delete data both from memory and from
the database. The `delete_trackers_from_memory` is called for outdated data,
given the database cleanup is delegated to the Gatekeeper, who will perform
a cascade cleanup on deleting the user's records. `delete_trackers` is called
for completed appointments which are completely wiped from the system.

The same split is used in the Watcher, only this time `delete_appointments_from_memory`
apply both to outdated data and data passed to (and accepted by) the Responder.
The former follow the same logic as with the Responder's case, as with the latter,
accepted appointment cleanup in the database is delegated to the Responder.