Bootstraps components using data from the DB.
- Repurposes how last_known_blocks are stored in the DB. Instead of each
component storing it's own last_known_block after block_connected, the
ChainMonitor is in charge of doing so once all of them have been updated.
- Given the previous change, block_connected needs to be safe to called
twice in two different execution cycles, since the system may crash before
all the components are updated (that was already an issue anyway when each
component was in charge of storing it's onw last_known_block).
- Splits the system initialization in three steps (main.rs):
    - First the components are created and the data is loaded from the DB
    - Second the components are brought up to date using block data that
    may have been missed while the system was offline
    - Last, the interfaces are brought up
- The ChainMonitor functionality is split so components can be brought
up to date before starting monitoring.
- Also fixes a bug with the LocatorCache where blocs where not being stored
in descending order (meaning the they were not evicted as expected)