Reformats Watcher to make it easier to test
Leaves an edge on bitcoind_unreachable unhandled atm:

If `Watcher::store_triggered_appointment` is called while bitcoind is unreachable the call will hang,
the request will timeout but be accepted. However, the user will not be handed the receipt.
This could be fixed adding a thread to take care of storing while the main thread returns the receipt.

Not fixing this atm since working with threads that call self.method is surprisingly non-trivial.