Revamps the Responder confirmation tracking
Confirmation count is not queried to bitcoind anymore. Instead, a height field is added to trackers and checked every new block.

The rationale behind this is modeling the confirmation tracking in a way that is more aligned with the block_connected / block_disconnected
approach followed by `chain::Listen`, so reorgs are easier to tackle. Also, this should optimize the confirmation tracking process:

Old:
getrawtransaction + parsing for each tracker every block_connected

New:
4 additional bytes per tracker + comparison (w/ subtraction) for each tracker every block_connected.