Adds dev-watchtower-max-retry-interval option to CLN plugin
`dev-watchtower-max-retry-interval` allows for fine tunning the waiting time
between intervals of the retrier. This makes it easier to test when the retrier
must have been attempted to send data to the tower again.

For now this is set as as a dev option. Not that matter much apart from it's naming.
(i.e. it can be set in any context). If it make sense for this fine tunning to be
offered to the user, the `dev-` prefix could be dropped and it could be documented.