fix: Fixes connection between gRPC client and server
In some rare occasions, when restarting the tower, the gRPC client is unable
to connect to the server. This is a temporary error that get's fixed by simply
waiting a bit. Looks like the root of the issue if the server not being fully up
when the client tries connect.

More info: https://github.com/talaia-labs/rust-teos/pull/65#issuecomment-1164469285

The fix mainly consists on  looping with a wait until the client is able to connect.