Adds a ready signal for the http API
In normal conditions, if the http API cannot bootstrap (due to the gRPC
server being temporary unavailable) it will just retry until it connects.

However, this is done asynchronously, meaning that it could be the case that
the tower is reported as ready when it is actually not. This can be hit during
E2E testing if the tower is restarted too quickly, resulting in a test failure:

https://github.com/sr-gi/rust-teos/runs/8074272734?check_suite_focus=true#step:9:5543
https://github.com/sr-gi/rust-teos/runs/8074272734?check_suite_focus=true#step:9:6970