fix: Arc retriers so we reduce the cloning + antipatern
One thing I really disliked about #89 was having a method that cloned its caller to be able to work around spawning a task inside it that called a method of the same class.

Turns out you can have self as Arc<Self> which would completely prevent having to do such a thing, plus it also reduces the number of things being cloned.