Update systemd,md
I have make some changes to the service. 

It's better to create it /etc/systemd/system and let /usr/lib/systemd/system for programs installed from repositories or .deb

Modified [unit] part, only bitcoind is necessary to be running but we want network too.

Modified [service] part. --datadir is not needed, if we add user and group the datadir will be the default folder /home/<USER>/.teos/

Deleted StandardOutput=journal and StandardError=journal and SyslogIdentifier, we have enough info without it. I think it's not necessary.

If the type is "simple" pidfile is not necessary

Added timeout and restartsec, I've put 60 seconds in restartsec because bitcoind needs some time to fully start.

Added some Hardening measures.