Patches watchtower-plugin test_get_appointment e2e
test_get_appointment fails when the plugin is asked about a given tracker before the
tower is able to process the appointment to tracker transition. This doesn't happen consistently,
but we've seen the error popping up every now and then.

Simply add a log based waiting to make sure the state transition is performed before asserting.