Adds time-waiting constants to retrier and redefine waiting times in tests
`retrier::tests::test_manage_retry_while_idle` was randomly failing (for Ubuntu) when checking whether the Retrier was
idle after giving up on a retry.

Some of the tests success is based on waiting time (given we need the retriers to either succeed or fail). Re-define the
waits so they are more accurate. Unfortunately there is no way of perfectly doing this, but the current setup seems not to fail