make stopper non-optional in tests
In some tests the carrier didn't need a stopper since it was already in the test scope so it would drop at the end of the test and do usual cleanup.
This commit unifies stuff and still moves the stopper inside the carrier so that all the tests look alike (not having ones with Some(stopper) and ones with None).